Track3/Backend/requirements.txt
# FastAPI & Core
fastapi
uvicorn[standard]
python-multipart

# Database
sqlalchemy
psycopg2-binary

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]

# Data Validation
pydantic
pydantic-settings
pydantic[email]

# HTTP Client & External APIs
httpx
requests

# Data Processing & Analytics
pandas
numpy
scipy

# Admin Interface
fastapi-amis-admin[postgresql]
asyncpg

# Development & Testing
pytest
pytest-asyncio
python-dotenv